backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
strawberry-graphql[fastapi]==0.246.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
celery==5.4.0
redis==5.0.8
chromadb==0.5.11
sentence-transformers==3.1.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
supabase==2.8.1
loguru==0.7.2
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.14.0
pytest==8.3.3
pytest-asyncio==0.24.0
